Moved the tests to own folder
